Adopt NSWDS
Adopt the HCS variant of the NSW Design System, revamp homepage to be more functional rather than promotional & start work on a revamped Start Here page that aims to promote the server's highlights.

Deleted the forms and links and FAQ pages as they are no longer necessary. Forms and links have been moved to the homepage whilst the FAQ has been merged with the new player guide (which is yet to be published on the website).